Split out component
