+++ ui/Screen/IdentityCreation.svelte
@@ -6,10 +6,10 @@
   import * as path from "../src/path.ts";
   import * as session from "../src/session.ts";
 
-  import { ModalLayout, RadicleLogo } from "../DesignSystem/Component";
-  import { Button } from "../DesignSystem/Primitive";
+  import { ModalLayout } from "../DesignSystem/Component";
 
-  import Form from "./IdentityCreation/Form.svelte";
+  import EnterName from "./IdentityCreation/EnterName.svelte";
+  import Welcome from "./IdentityCreation/Welcome.svelte";
   import Success from "./IdentityCreation/Success.svelte";
 
   const returnToWelcomeStep = () => {
@@ -42,7 +42,7 @@
     switch ($store) {
       case State.Welcome:
         return;
-      case State.Form:
+      case State.EnterName:
         returnToWelcomeStep();
         return;
       case State.SuccessView:
@@ -57,39 +57,11 @@
   };
 </script>
 
-<style>
-  .container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    height: 100%;
-  }
-
-  .container p {
-    text-align: center;
-    color: var(--color-foreground-level-5);
-    margin: 2.5rem 0;
-    max-width: 20rem;
-  }
-</style>
-
 <ModalLayout escapable={false} {onClose}>
   {#if $store === State.Welcome}
-    <div class="container">
-      <RadicleLogo />
-      <p>
-        A free and open-source way to host, share, and build software together.
-      </p>
-      <Button
-        style="flex-shrink: 0; margin-bottom: 24px;"
-        on:click={() => store.set(State.Form)}
-        dataCy="get-started-button">
-        Get started
-      </Button>
-    </div>
-  {:else if $store === State.Form}
-    <Form
+    <Welcome on:next={() => store.set(State.EnterName)} />
+  {:else if $store === State.EnterName}
+    <EnterName
       on:cancel={returnToWelcomeStep}
       on:error={onError}
       on:success={() => store.set(State.SuccessView)} />
